+++ lib/models/Installation.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/foundation.dart';
+
 enum OS { Android, iOS }
 
 enum PushServiceType { GCM, Firebase, APNS }
@@ -6,10 +8,10 @@ class Installation {
   final String? pushRegistrationId;
   final String? pushServiceToken;
   final PushServiceType? pushServiceType;
-  final bool? isPrimaryDevice;
-  final bool? isPushRegistrationEnabled;
-  final bool? notificationsEnabled;
-  final bool? geoEnabled;
+  bool? isPrimaryDevice;
+  bool? isPushRegistrationEnabled;
+  bool? notificationsEnabled;
+  bool? geoEnabled;
   final String? sdkVersion;
   final String? appVersion;
   final OS? os;
@@ -21,28 +23,28 @@ class Installation {
   final String? deviceTimezoneOffset;
   final String? applicationUserId;
   final String? deviceName;
-  final Map<String, dynamic>? customAttributes;
+  Map<String, dynamic>? customAttributes;
 
   Installation(
       {this.pushRegistrationId,
-        this.pushServiceType,
-        this.pushServiceToken,
-        this.isPrimaryDevice,
-        this.isPushRegistrationEnabled,
-        this.notificationsEnabled,
-        this.geoEnabled,
-        this.sdkVersion,
-        this.appVersion,
-        this.os,
-        this.osVersion,
-        this.deviceManufacturer,
-        this.deviceModel,
-        this.deviceSecure,
-        this.language,
-        this.deviceTimezoneOffset,
-        this.applicationUserId,
-        this.deviceName,
-        this.customAttributes});
+      this.pushServiceType,
+      this.pushServiceToken,
+      this.isPrimaryDevice,
+      this.isPushRegistrationEnabled,
+      this.notificationsEnabled,
+      this.geoEnabled,
+      this.sdkVersion,
+      this.appVersion,
+      this.os,
+      this.osVersion,
+      this.deviceManufacturer,
+      this.deviceModel,
+      this.deviceSecure,
+      this.language,
+      this.deviceTimezoneOffset,
+      this.applicationUserId,
+      this.deviceName,
+      this.customAttributes});
 
   static PushServiceType? resolvePushServiceType(String? pst) {
     if (pst == null) {
@@ -68,32 +70,34 @@ class Installation {
   }
 
   Map<String, dynamic> toJson() => {
-    'pushRegistrationId': pushRegistrationId,
-    'pushServiceToken': pushServiceToken,
-    'pushServiceType': pushServiceType,
-    'isPrimaryDevice': isPrimaryDevice,
-    'isPushRegistrationEnabled': isPushRegistrationEnabled,
-    'notificationsEnabled': notificationsEnabled,
-    'geoEnabled': geoEnabled,
-    'sdkVersion': sdkVersion,
-    'appVersion': appVersion,
-    'os': os,
-    'osVersion': osVersion,
-    'deviceManufacturer': deviceManufacturer,
-    'deviceModel': deviceModel,
-    'deviceSecure': deviceSecure,
-    'language': language,
-    'deviceTimezoneOffset': deviceTimezoneOffset,
-    'applicationUserId': applicationUserId,
-    'deviceName': deviceName,
-    'customAttributes': customAttributes,
-  };
+        'pushRegistrationId': pushRegistrationId,
+        'pushServiceToken': pushServiceToken,
+        'pushServiceType':
+            pushServiceType != null ? describeEnum(pushServiceType!) : null,
+        'isPrimaryDevice': isPrimaryDevice,
+        'isPushRegistrationEnabled': isPushRegistrationEnabled,
+        'notificationsEnabled': notificationsEnabled,
+        'geoEnabled': geoEnabled,
+        'sdkVersion': sdkVersion,
+        'appVersion': appVersion,
+        'os': os != null ? describeEnum(os!) : null,
+        'osVersion': osVersion,
+        'deviceManufacturer': deviceManufacturer,
+        'deviceModel': deviceModel,
+        'deviceSecure': deviceSecure,
+        'language': language,
+        'deviceTimezoneOffset': deviceTimezoneOffset,
+        'applicationUserId': applicationUserId,
+        'deviceName': deviceName,
+        'customAttributes': customAttributes,
+      };
 
   Installation.fromJson(Map<String, dynamic> json)
       : isPrimaryDevice = json['isPrimaryDevice'],
         pushRegistrationId = json['pushRegistrationId'],
         pushServiceToken = json['pushServiceToken'],
-        pushServiceType = Installation.resolvePushServiceType(json['pushServiceType']),
+        pushServiceType =
+            Installation.resolvePushServiceType(json['pushServiceType']),
         isPushRegistrationEnabled = json['isPushRegistrationEnabled'],
         notificationsEnabled = json['notificationsEnabled'],
         geoEnabled = json['geoEnabled'],
@@ -110,7 +114,7 @@ class Installation {
         deviceName = json['deviceName'],
         customAttributes = json['customAttributes'];
 
-  String? getPushRegistrationId () {
+  String? getPushRegistrationId() {
     return this.pushRegistrationId;
   }
 }